seprated the routes
