+++ mypofo/app.js
@@ -3,7 +3,10 @@ const hbs = require("hbs");
 const session = require("express-session");
 const mongoose = require("mongoose");
 const middleware = require("./middlewares/middleware");
-const routes = require("./routes");
+const index = require("./routes/index");
+const projectRoutes = require("./routes/projects");
+const blogRoutes = require("./routes/blogs");
+const adminRoutes = require("./routes/admin");
 const app = express();
 
 app.set("view engine", "hbs");
@@ -34,22 +37,10 @@ app.use(
 app.use(express.static(__dirname + "/static"));
 app.use(middleware.authenticated);
 
-app.get("/", routes.index);
-app.get("/projects", routes.projectList);
-app.get("/blogs", routes.blogs);
-app.get("/contact", routes.contact);
-app.get("/signin", routes.signIn);
-app.post("/signin", routes.doSignin);
-app.get("/signup", routes.signUp);
-app.get("/logout", routes.logout);
-app.post("/projects", routes.createProject);
-
-app.get("/projects/:alias", middleware.authenticate, routes.projectDetail);
-app.get("/blogs/:alias", middleware.authenticate, routes.blogDetail);
-
-// admin routes
-app.get("/admin", middleware.authenticate, routes.admin);
-app.get("/admin/projects", middleware.authenticate, routes.adminProjectList);
+app.use("/", index);
+app.use("/projects", projectRoutes);
+app.use("/blogs", blogRoutes);
+app.use("/admin", middleware.authenticate, adminRoutes);
 
 app.use(middleware.notFound);
 app.use(middleware.handleError);
