+++ mypofo/middlewares/middleware.js
@@ -1,4 +1,4 @@
-module.exports.logger = (req, res, next) => {
+exports.logger = (req, res, next) => {
   console.log(req.method, req.url);
   next();
 };
