+++ mypofo/routes/index.js
@@ -1,101 +1,39 @@
 const data = require("./data").data;
-const ProjectService = require("../services/projectService");
+const router = require("express").Router();
 
 exports.createProject = (req, res) => {
   let data = req.body;
   ProjectService.create(data);
 };
 
-exports.index = (req, res) => {
+router.get("/", (req, res) => {
   res.render("home", {
     title: "Abhishek - Portfolio",
     layout: "homeLayout",
   });
-};
+});
 
-exports.projectList = (req, res, next) => {
-  function callback(err, dt) {
-    if (err) {
-      console.log(err);
-      next(err);
-    } else {
-      res.render("projects", {
-        title: "Projects",
-        navproject: true,
-        layout: "layout",
-        projects: dt,
-      });
-    }
-  }
-  ProjectService.list(callback);
-};
-
-exports.blogs = (req, res) => {
-  res.render("blogs", {
-    title: "Blogs",
-    layout: "layout",
-    navBlogs: true,
-  });
-};
-
-exports.contact = (req, res) => {
+router.get("/contact", (req, res) => {
   res.render("contact", {
     title: "contact",
     layout: "layout",
     navContact: true,
   });
-};
+});
 
-exports.projectDetail = (req, res) => {
-  const alias = req.params.alias;
-
-  let dt = data.projects[data.projectIndex[alias]];
-
-  res.render("projectDetail", {
-    title: "Project Detail",
-    layout: "layout",
-    project: dt,
-  });
-};
-
-exports.blogDetail = (req, res) => {
-  let params = req.params.alias;
-  let title = params
-    .split("-")
-    .map((e) => e.slice(0, 1).toUpperCase() + e.slice(1))
-    .join(" ");
-  res.render("blogDetail", {
-    title: `Blog - ${title}`,
-    layout: "layout",
-  });
-};
-
-exports.signIn = (req, res) => {
+router.get("/signin", (req, res) => {
   res.render("signin", {
     title: "signin",
     layout: "loginLayout",
   });
-};
+});
 
-exports.signUp = (req, res) => {
+router.get("/signup", (req, res) => {
   res.render("signup", {
     title: "Signup",
     layout: "loginlayout",
   });
-};
-
-exports.admin = (req, res) => {
-  res.render("admin/index", {
-    layout: "adminLayout",
-  });
-};
-
-exports.adminProjectList = (req, res) => {
-  res.render("admin/projectList", {
-    layout: "adminLayout",
-    projects: data.projects,
-  });
-};
+});
 
 let user = [
   {
@@ -115,7 +53,7 @@ let user = [
   },
 ];
 
-exports.doSignin = (req, res) => {
+router.post("/signin", (req, res) => {
   const data = req.body;
   let findUser = user.filter((e) => e.email == data.email);
   if (findUser.length > 0) {
@@ -137,10 +75,12 @@ exports.doSignin = (req, res) => {
       message: "Email or password is incorrect",
     });
   }
-};
+});
 
-exports.logout = (req, res) => {
+router.get("/logout", (req, res) => {
   req.session.isLoggedIn = false;
   delete res.locals.user;
   res.redirect("/");
-};
+});
+
+module.exports = router;
